Adjusted code to fit analyzer rules
