+++ src/NoPlan.Api/Extensions/ClaimsPrincipalExtensions.cs
@@ -16,5 +16,5 @@ public static class ClaimsPrincipalExtensions
     public static Guid GetId(this ClaimsPrincipal user) =>
         Guid.TryParse(user.GetObjectId(), out var id)
             ? id
-            : default;
+            : Guid.Empty;
 }
