+++ src/NoPlan.Contracts/Requests/V1/ToDos/UpdateTagRequest.cs
@@ -3,6 +3,7 @@
 public sealed record UpdateTagRequest
 {
     public Guid Id { get; init; }
+
     public string Name { get; init; } = null!;
 }
