+++ src/NoPlan.Contracts/Responses/V1/ToDos/TagResponse.cs
@@ -3,6 +3,8 @@
 public sealed record TagResponse
 {
     public Guid Id { get; set; }
+
     public string Name { get; set; } = null!;
+
     public DateTime AssignedAt { get; set; }
 }
