+++ src/NoPlan.Infrastructure/Data/EntityTypeConfigurations/ToDoEntityTypeConfiguration.cs
@@ -3,7 +3,7 @@ using NoPlan.Infrastructure.Data.Models;
 
 namespace NoPlan.Infrastructure.Data.EntityTypeConfigurations;
 
-public sealed class ToDoEntityTypeConfiguration : IEntityTypeConfiguration<ToDo>
+internal sealed class ToDoEntityTypeConfiguration : IEntityTypeConfiguration<ToDo>
 {
     public void Configure(EntityTypeBuilder<ToDo> builder) =>
         builder.OwnsMany(t => t.Tags);
