+++ src/NoPlan.Infrastructure/Data/Models/Tag.cs
@@ -3,6 +3,8 @@
 public sealed class Tag
 {
     public Guid Id { get; set; }
+
     public string Name { get; set; } = null!;
+
     public DateTime AssignedAt { get; set; }
 }
