+++ src/NoPlan.Infrastructure/Options/AppConfigurationOptions.cs
@@ -7,26 +7,26 @@ namespace NoPlan.Infrastructure.Options;
 /// </summary>
 public sealed class AppConfigurationOptions : IOptionsSectionDefinition
 {
+    /// <inheritdoc cref="IOptionsSectionDefinition.SectionName" />
+    public static string SectionName => "AppConfiguration";
+
     /// <summary>
-    ///     The <see cref="Uri" /> of the Azure App Configuration endpoint.
+    ///     Gets or sets the <see cref="Uri" /> of the Azure App Configuration endpoint.
     /// </summary>
     public Uri EndPoint { get; set; } = null!;
 
     /// <summary>
-    ///     The Azure Service Bus namespace.
+    ///     Gets or sets the Azure Service Bus namespace.
     /// </summary>
     public string ServiceBusNamespace { get; set; } = null!;
 
     /// <summary>
-    ///     The name of the Azure Service Bus topic receiving App Configuration events.
+    ///     Gets or sets the name of the Azure Service Bus topic receiving App Configuration events.
     /// </summary>
     public string ServiceBusTopicName { get; set; } = null!;
 
     /// <summary>
-    ///     The name of the Azure Service Bus subscription receiving App Configuration events.
+    ///     Gets or sets the name of the Azure Service Bus subscription receiving App Configuration events.
     /// </summary>
     public string ServiceBusSubscriptionName { get; set; } = null!;
-
-    /// <inheritdoc />
-    public static string SectionName => "AppConfiguration";
 }
